requirements.txt
streamlit==1.39.0
pandas>=2.2.2
python-dateutil>=2.9.0
pytz>=2024.1
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
feedparser>=6.0.11
rapidfuzz>=3.9.6
folium>=0.17.0
icalendar>=5.0.12